+++ core/src/main/java/com/galinazabelina/core/core/repository/DebitAccountRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
@@ -15,9 +16,9 @@ public interface DebitAccountRepository extends CrudRepository<Account, Long> {
     void closeAccountById(@Param("accountId") Long accountId);
 
     @Query("select balance from Account where id = :accountId")
-    Optional<Long> getBalance(@Param("accountId") Long accountId);
+    Optional<BigDecimal> getBalance(@Param("accountId") Long accountId);
 
     @Query("update Account set balance = balance + :balanceChange where id = :accountId")
-    void changeBalance(@Param("accountId") Long accountId, @Param("balanceChange") Long balanceChange);
+    void changeBalance(@Param("accountId") Long accountId, @Param("balanceChange") BigDecimal balanceChange);
 
 }
